requirements.txt
streamlit
sqlalchemy
pandas
python-dotenv
matplotlib